Подкатегории

+++ admin.php
@@ -50,6 +50,18 @@ switch ($action) {
     case 'deleteUser':
         deleteUser();
         break;
+    case 'deleteSubcategory':
+        deleteSubcategory();
+        break;
+    case 'newSubcategory':
+        newSubcategory();
+        break;
+    case 'editSubcategory':
+        editSubcategory();
+        break;
+    case 'listSubcategories':
+        listSubcategories();
+        break;
     default:
         listArticles();
 }
@@ -106,23 +118,21 @@ function newArticle() {
     $results = array();
     $results['pageTitle'] = "New Article";
     $results['formAction'] = "newArticle";
+    $results['group'] = Subcategory::getGroup();
+    $results['categories'] = Category::getList()['results'];
+    $results['subcategories'] = Subcategory::getList()['results'];
 
-    if ( isset( $_POST['saveChanges'] ) ) {
-//            echo "<pre>";
-//            print_r($results);
-//            print_r($_POST);
-//            echo "<pre>";
-//            В $_POST данные о статье сохраняются корректно
-        // Пользователь получает форму редактирования статьи: сохраняем новую статью
-        $article = new Article();
-        $article->storeFormValues( $_POST );
-//            echo "<pre>";
-//            print_r($article);
-//            echo "<pre>";
-//            А здесь данные массива $article уже неполные(есть только Число от даты, категория и полный текст статьи)          
-        $article->insert();
-        header( "Location: admin.php?status=changesSaved" );
-
+    if (isset($_POST['saveChanges'])) {
+	$article = new Article();
+        $article->storeFormValues($_POST);
+	if ($_POST['categoryId'] && $_POST['subcategoryId'] && $results['group'][$_POST['categoryId']] !== []
+		&& !in_array($_POST['subcategoryId'], $results['group'][$_POST['categoryId']])) {
+	    $results['errorMessage'] = 'Ошибка: категория не соответствует подкатегории';
+	    require (TEMPLATE_PATH . '/admin/editArticle.php');
+	} else {
+	    $article->insert();
+	    header("Location: admin.php?status=changesSaved");
+	}
     } elseif ( isset( $_POST['cancel'] ) ) {
 
         // Пользователь сбросил результаты редактирования: возвращаемся к списку статей
@@ -131,8 +141,8 @@ function newArticle() {
 
         // Пользователь еще не получил форму редактирования: выводим форму
         $results['article'] = new Article;
-        $data = Category::getList();
-        $results['categories'] = $data['results'];
+//        $data = Category::getList();
+//        $results['categories'] = $data['results'];
         require( TEMPLATE_PATH . "/admin/editArticle.php" ); 
     }
 }
@@ -148,6 +158,9 @@ function editArticle() {
     $results = array();
     $results['pageTitle'] = "Edit Article";
     $results['formAction'] = "editArticle";
+    $results['group'] = Subcategory::getGroup();
+    $results['categories'] = Category::getList()['results'];
+    $results['subcategories'] = Subcategory::getList()['results'];
 
     if (isset($_POST['saveChanges'])) {
 
@@ -156,10 +169,17 @@ function editArticle() {
             header( "Location: admin.php?error=articleNotFound" );
             return;
         }
-
-        $article->storeFormValues( $_POST );
-        $article->update();
-        header( "Location: admin.php?status=changesSaved" );
+        
+        $article->storeFormValues($_POST);
+        
+	if ($_POST['categoryId'] && $_POST['subcategoryId'] && $results['group'][$_POST['categoryId']] !== []
+		&& !in_array($_POST['subcategoryId'], $results['group'][$_POST['categoryId']])) {
+	    $results['errorMessage'] = 'Ошибка: категория не соответствует подкатегории';
+	    require (TEMPLATE_PATH . '/admin/editArticle.php');
+	} else {
+	    $article->update();
+	    header("Location: admin.php?status=changesSaved");
+	}
 
     } elseif ( isset( $_POST['cancel'] ) ) {
 
@@ -169,8 +189,8 @@ function editArticle() {
 
         // Пользвоатель еще не получил форму редактирования: выводим форму
         $results['article'] = Article::getById((int)$_GET['articleId']);
-        $data = Category::getList();
-        $results['categories'] = $data['results'];
+//        $data = Category::getList();
+//        $results['categories'] = $data['results'];
         require(TEMPLATE_PATH . "/admin/editArticle.php");
     }
 
@@ -197,13 +217,8 @@ function listArticles() {
     $data = Article::getList();
     $results['articles'] = $data['results'];
     $results['totalRows'] = $data['totalRows'];
-    
-    $data = Category::getList();
-    $results['categories'] = array();
-    foreach ($data['results'] as $category) { 
-        $results['categories'][$category->id] = $category;
-    }
-    
+    $results['categories'] = Category::getList()['results'];
+    $results['subcategories'] = Subcategory::getList()['results'];
     $results['pageTitle'] = "Все статьи";
 
 
@@ -234,6 +249,9 @@ function listCategories() {
     if ( isset( $_GET['error'] ) ) {
         if ( $_GET['error'] == "categoryNotFound" ) $results['errorMessage'] = "Error: Category not found.";
         if ( $_GET['error'] == "categoryContainsArticles" ) $results['errorMessage'] = "Error: Category contains articles. Delete the articles, or assign them to another category, before deleting this category.";
+        if ( $_GET['error'] == "categoryContainsSubcategories" ) $results['errorMessage'] = "Ошибка: Категория содержит подкатегории. "
+		. "Удалите подкатегории или назначьте их другой категории перед удалением этой категории.";
+        
     }
 
     if ( isset( $_GET['status'] ) ) {
@@ -256,9 +274,14 @@ function newCategory() {
         // User has posted the category edit form: save the new category
         $category = new Category;
         $category->storeFormValues( $_POST );
-        $category->insert();
-        header( "Location: admin.php?action=listCategories&status=changesSaved" );
 
+        if (Subcategory::nameExist($category->name, true)) {
+	    $results['errorMessage'] = 'Ошибка: Категория существует';
+	    require (TEMPLATE_PATH . '/admin/editCategory.php');
+	} else {
+	    $category->insert();
+	    header("Location: admin.php?action=listCategories&status=changesSaved");
+	}
     } elseif ( isset( $_POST['cancel'] ) ) {
 
         // User has cancelled their edits: return to the category list
@@ -288,9 +311,15 @@ function editCategory() {
           return;
         }
 
-        $category->storeFormValues( $_POST );
-        $category->update();
-        header( "Location: admin.php?action=listCategories&status=changesSaved" );
+        $category->storeFormValues($_POST);
+        if (Subcategory::nameExist($category->name, true) && 
+		$category->id != Subcategory::nameId($category->name, true)) {
+	    $results['errorMessage'] = 'Ошибка: Категория существует';
+	    require(TEMPLATE_PATH . '/admin/editCategory.php');
+	} else {
+	    $category->update();
+	    header("Location: admin.php?action=listCategories&status=changesSaved");
+	}
 
     } elseif ( isset( $_POST['cancel'] ) ) {
 
@@ -313,13 +342,19 @@ function deleteCategory() {
         return;
     }
 
-    $articles = Article::getList( 1000000, $category->id );
+    $articles = Article::getList( 1000000, $category->id, 'publicationDate ASC', true );
 
     if ( $articles['totalRows'] > 0 ) {
         header( "Location: admin.php?action=listCategories&error=categoryContainsArticles" );
         return;
     }
 
+    $subcategories = Subcategory::getList(1000000, 'name ASC', $category->id);
+    if (count($subcategories['results']) > 0) {
+        header("Location: admin.php?action=listCategories&error=categoryContainsSubcategories");
+        return;
+    }
+    
     $category->delete();
     header( "Location: admin.php?action=listCategories&status=categoryDeleted" );
 }
@@ -399,4 +434,102 @@ function deleteUser() {
     }
     $user->delete();
     header("Location: admin.php?action=listUsers&status=userDeleted");
-}        
+} 
+
+function newSubcategory() {
+
+    $results = array();
+    $results['pageTitle'] = "Новая подкатегория";
+    $results['formAction'] = "newSubcategory";
+    $data = Category::getList();
+    $results['categories'] = $data['results'];
+    if (isset($_POST['saveChanges'])) {
+        $subcategory = new Subcategory;
+        $subcategory->storeFormValues($_POST);
+	if (Subcategory::nameExist($subcategory->name)) {
+	    $results['errorMessage'] = 'Ошибка: Подкатегория существует';
+	    require (TEMPLATE_PATH . '/admin/editSubcategory.php');
+	} else {
+	    $subcategory->insert();
+	    header("Location: admin.php?action=listSubcategories&status=changesSaved");
+	}
+    } elseif (isset($_POST['cancel'])) {
+        // Пользователь отменяет добавление: возврат к списку подкатегорий
+        header("Location: admin.php?action=listSubcategories");
+    } else {
+	// Пользователь еще не получил форму редактирования: выводим форму
+        $results['subcategory'] = new Subcategory;
+
+	require(TEMPLATE_PATH . "/admin/editSubcategory.php");
+    }
+}
+
+function editSubcategory() {
+
+    $results = array();
+    $results['pageTitle'] = "Редактировать подкатегорию";
+    $results['formAction'] = "editSubcategory";
+    $data = Category::getList();
+    $results['categories'] = $data['results'];
+
+    if (isset($_POST['saveChanges'])) {
+	$subcategory = Subcategory::getById((int) $_POST['subcategoryId']);
+	$subcategory->storeFormValues($_POST);
+	if (Subcategory::nameExist($subcategory->name) && 
+		$subcategory->id != Subcategory::nameId($subcategory->name)) {
+	    $results['errorMessage'] = 'Ошибка: Подкатегория существует';
+	    require(TEMPLATE_PATH . '/admin/editSubcategory.php');
+	} else {
+	    $subcategory->update();
+	    header("Location: admin.php?action=listSubcategories&status=changesSaved");
+	}
+    } elseif (isset($_POST['cancel'])) {
+        // Пользователь отменяет редактирование: возврат к списку подкатегорий
+        header("Location: admin.php?action=listSubcategories" );
+    } else {
+        // Отобразить данные подкатегории перед редактированием
+        $results['subcategory'] = Subcategory::getById((int) $_GET['subcategoryId']);
+	require(TEMPLATE_PATH . "/admin/editSubcategory.php");
+    }
+}
+
+function listSubcategories() {
+
+    $results = array();
+    $data = Subcategory::getList();
+    $results['subcategories'] = $data['results'];
+    $results['totalRows'] = $data['totalRows'];
+    $data = Category::getList();
+    $results['categories'] = array();
+    foreach ($data['results'] as $category) { 
+        $results['categories'][$category->id] = $category;
+    }
+    $results['pageTitle'] = "Подкатегории";
+    if (isset($_GET['error'])) {
+	if ($_GET['error'] == "subcategoryContainsArticles") $results['errorMessage'] = "Ошибка: Подкатегория содержит статьи. "
+		. "Удалите статьи или назначьте их другой подкатегории перед удалением этой подкатегории.";
+    }
+    if (isset($_GET['status'])) {
+        if ($_GET['status'] == "changesSaved") {
+	    $results['statusMessage'] = "Изменения сохранены.";
+	} elseif ($_GET['status'] == "subcategoryDeleted") {
+	    $results['statusMessage'] = "Подкатегория удалена.";
+	}
+    }
+    require(TEMPLATE_PATH . "/admin/listSubcategories.php");
+}
+
+function deleteSubcategory() {
+
+    if (!$subcategory = Subcategory::getById((int) $_GET['subcategoryId'])) {
+        header("Location: admin.php?action=listSubcategories&error=subcategoryNotFound");
+        return;
+    }
+    $articles = Article::getBySubcat(1000000, $subcategory->id, true, 'publicationDate ASC');
+    if ($articles['totalRows'] > 0) {
+        header("Location: admin.php?action=listSubcategories&error=subcategoryContainsArticles" );
+        return;
+    }
+    $subcategory->delete();
+    header("Location: admin.php?action=listSubcategories&status=subcategoryDeleted");
+}
